Restrict engines:node to >=12, <15, Update apollo-server-core to 2.25.3
This brings `apollo-server-core` up to the latest released version and restricts this version of the integration to have the same Node engine support as the 2.x branch of Apollo.

This is due to the fact that batch requests no longer run in parallel in Node 16 and the code that handles that is outside the scope of this package. If the Apollo team adds Node 16 support to the v2 branch, then I can update dependencies to enable support as well, but until then I can't do anything about it.

Vercel at this time only supports two environments anyways, Node 12 and 14.

Node 16 support will be added in a new major version, along with an update to Apollo Server 3.x